Fix Block serialization (Miners).
We were adding the nonce in when parsing, despite the fact the nonce is right before the miners anyways.

This meant in Serialize, we had to subtract a constant from the length.

We no longer suntract a constant. We no longer add the nonce in ourselves in Parse.

The tests have also been edited, to prove our fix works, to use a random nonce, with a length of anywhere from 1 to 3 bytes (max of 100k). I also made the proof random while I was in there.