Store Verifications in the DB, and load archived Verifications when we start the node.
This is a lot of work that needs more thorough testing. Thanks to the Epoch Tests, the basic functionality has been tested, but I'm concerned there's an index error SOMEWHERE (whether it be an off by one when accessing the Verifications, or accessing the in-memory unarchived Verifications when we wanted to access all of them).

This also removed the archived field from Verification, as that data wasn't being preserved when serialized and we can rescan the chain if we need it.

Finally, I have explicitly added to the TODO that we should change archived (Verifier) to an int so we have -1, but since we can get the start of the unarchived Verifications via verifier.verifications[0], this is way less of a problem