Tweaks.
Move Serialize's strip to a dedicated function in Util.

Update our strip algo (new one is roughly 3x faster).

Make sure verifiers added to the verifiersStr are padded.

Make sure verifiers grabbed from the verifiersStr are stripped.

Add a verifiersSeq so we can double check if a verifier is already in the verifiersStr.

Double check if the verifier is already in the verifiersStr.

Don't only set C++ if the user said to use the C backend. No other language option will work.