Update Lattice to not clear the cache once an Entry is verified, but rather once an Entry is out of the Epochs.
This fixes a bug, introduced when we added the DB, which broke the blockchain. Basically, we were removing items from the lookup when they were verified, but adding the Block with said verifications requires the lookup.

We could have accessed the database, but as we move into tightening the protocol, we need a more proper solution (the one implemented). account.confirmed now represents the amount of Entries out of an Epoch and assumes no gaps, instead of being the highest confirmed Entry with no gaps below it.

This was not caught due to a lack of integration tests.

This should not have been fixed on this branch.

This bug must be fixed for this branch to be completed.

The related bugs were initially thought to be much smaller in scope, and therefore the first set of fixes were put on this branch. There's no reason to split them up now.