Both MainMerit and NetworkSync verify the aggregate signature; move this code out to Block.nim.
It should be possible to remove this code entirely from MainMerit, as sync will throw an error if it fails to get everything, and sync verifies the signature.

That said, I'm not comfortable to fully remove it yet. The flow when we receive a Block, instead of when we ask for one, needs to be double checked