Clean the TODO.
Move "Clean tests." out of the cleanup as the tests are fully functional and don't affect the runtime.

Remove "General Cleanup" as the only other thing needing general cleanup was the tests.

Remove merging "verifications.getPendingAggregate" with "vverification.getUnarchivedIndexes". We'd need a new data type and the RPC already combines them.

Move "Don't rebroadcast data that we're syncing." and "Pass difficulties to the parsing functions to immediately check if work was put into a Block/Entry (stop DoS attacks)." to Network as those aren't cleanups but optimizations.

Remove "Remove the code from MainMerit..." as that's done.

Move the tests/ back to tests as those aren't cleaning, they're tests.