Remove fcBoundsOverride.
Instead of being bundled with forceCheck, `boundsCheck` is now a dedicated pragma

We could update to it, but now that we've confirmed we handle bounds properly, albeit relying on Nim's bounds checks, there's no need. We just need to pay more attention to make sure this isn't a problem in the future.