Replace GMP with the faster StInt.
GMP had a package/dll dependency.

Our GMP package didn't destruct its numbers, so it had a memory leak.

The StInt benchmark, when modified to also test GMP, proves StInt is faster.

We didn't use it historically as its numbers are fixed size. Now that we're only using it for difficulty calculations, there is a fixed size; 384 bytes. Unfortunately, StInt requires numbers be a power of 2, so we must use the 512 bit size, but difficulty calculations happen every 10 minutes at max and we no longer have a memory leak, so it's tolerable.