Add in Potential Pending Debt.
Six blocks in, on the fourth call to test(), the original Lattice has the potential debt as 0, but the reloaded thinks it's still 150.

We're reloading an Entry from RAM but not its Verifications.

The two places were this is likely to occur:
Lattice.nim's lines 221-ish (blockchain.height - 5), yet I updated that to 6. Unfortunately, that didn't fix the bug, but it didn't make things worse either.

newAccountObj's loading from confirmed ..< height. I'm pretty sure confirmed is includive, meaning we're loading a confirmed Entry.

Honestly, abstracting the DB to take in obects, and having a dedicated set of Serialize functions is a good start. Beyond that, we need to remove the 500 line Lattice file we have,