Fix https://github.com/MerosCrypto/Meros/issues/55 (see description).
We already 'fixed' this. It turns out then I deleted the comment about why the check was commented, not the comment mark on the check,

When I fixed that, it failed. Looking into why, it turns out we reclaculated the weight using the same weight for every Verifier (whatever the weight of the newest Verification was).

That is, of course, a major problem. This corrects all checks, simplifies the existing algorithm, and fixes it.

Thanks to the new weights table, we can now jnot ust confirm that both agree on what is verified, but also that both agree on the amount of Merit behind it.