Unify and reorganize the Wallet libraries.
Wallet is now a Mnemonic + HDWallet. It generates Mnemonics until one creates a valid HDWallet. It also offers several procs to easily access the contained HDWallet.

This successfully updates tests, main, the RPC, and the GUI.

This also fixes a couple bugs discovered in the process (invalid secret derivation due to spacing and invalid checksum calculation).