Add the initial Python code.
These include a BlockHeader class, partial BlockBody class, Block class, and Blockchain class.

The BlockBody is partial as it doesn't handle records/miners beyond the required miner which is randomly generated and inlined.

The Blockchain class simply keeps track of Blocks and calculates difficulty, as that's all it needs to do. This is a test suite, not a node.

A test, which doesn't start the node itself or work off vectors, is included. It advances the chain off the RPC and verifies the produced blocks are added.