Redesign the RPC.
Provides a nifty macro for declaring routes, and for merging multiple routers into one (so each module has a router).

The RPC no longer has a GlobalFunctionBox (the route constructor takes it as an argument), list of clients (this was never needed), or tracks if it's listening (if the socket exists, it's listening).