Updated addBlock for MeritRemovals.
addBlock now accepts records which involve the pending the MeritRemoval/removes Merit when a MeritHolder is archived.

addBlock does have the proper event flow for MeritRemoval (set rewards to 0, also remove rewards in that Block), yet Merit still calculates rewards with minted Merit.

This is on a new branch because it's going to break compilation.