Remove unknown hashes.
Technically, we still support unknown verifications. Unsigned verifications can be unknown, as we apply Elements before Transactions.

The Network verifies it has all the transactions before applying Elements. Since these Transactions are fully parsed, they are theoretically valid. If they fail to add, it's because a competing Transaction was verified. Even if they fail to add, we do still save them to the Database.

Since 'unknown' Verifications are only unknown while adding the Block, since we add every Transaction immediately after adding said Verifications, and we commit to the DB AFTER every Block addition, the DB doesn't need to know anything about 'unknown'.

The reason for saving transactions which were beat is too allow other node's to sync in the future (as they need proof the Verification isn't of an unknown hash). This does lower the amount of pruning we can do, yet the Transactions are a drop in the pond compared to the Elements.

As unknown hash Verifications are still valid causes for a MeritRemoval, we now need to check if a SignedVerification is malicious before trying to add it. This has been accomplished with checkMalicious.

This updates the MeritRemoval tests, which used Verifications with unknown hashes, to also include a Data Transaction, sent as needed.

The MeritRemoval generators also used Merit when they only needed Blockchain. That has been corrected.

Finally, this fixes bugs where Blocks with invalid records would cause the node to error out, instead of rejecting the Block.