Work with PendingActions.
Reverts pending actions once a MeritHolder has a MeritRemoval. Does not unverify Transactions, even if they should be. Does reapply pending actions if they should be reapplied.

Changes the threshold formula from having a 600 offset to a 1200 offset to count for imperfect reversions/reapplications caused in relation to MeritRemovals.

Adds Cause/Live tests for PendingActions, along with a generator and vectors.

Fixes incorrect MeritRemoval message lengths in Meros.py.

Fixes a bug in TDBTest where if a key has the amount it's going to spend, and its minted +0, the change output will have 0.