Start rebuilding the Transactions DAG and fix https://github.com/MerosCrypto/Meros/issues/84.
This is a huge work in progress. It completely removes all tracking of Consensus from the Transactions DAG, solely keeping Transactions/output tracking. There's no longer any Difficulties/weight tracking.

Why? The Difficulties were already planned to be moved to the Consensus DAG. Then, we can apply Difficulty votes without breaking encapsulation. Moving weights, and therefore whether or not a Transaction is verified, is the next logical step. The Transactions DAG now solely considers "Is this a valid Transaction?" based on the inputs, outputs, and signature.

The TDBTest was modified to only handle Mints, Claims, and Sends (no longer Datas). It does pass. It no longer attempts Datas as this commit breaks them by not tracking the relevant info needed to verify them. This commit also disables safe Verifications by not tracking output spends, so Merit Holders would verify competing Transactions.

This is on a branch for two reasons. The first is because it's a massive change that should be branched. The second is because it's a breaking change, as proven by the fact Datas don't currently work and main no longer compiles.

By the time this branch is merged, I aim to have the following things done:
- Restore Data functionality by saving their senders/tracking tips.
- Track input spends, firstly at all, and secondly in a way beneficial to MeritRemovals.

- Move Transaction Difficulties to the Consensus DAG and check the Transaction proofs when they're parsed.
- Move Transaction weights and verified/not verified to the Consensus DAG.

- Correctly "unverify" Transactions.
- Correct the Verification threshold.
- Reload Verifications with their MeritHolder's current Merit.

We've already fixed https://github.com/MerosCrypto/Meros/issues/84.