Don't finalize Parents before their Children.
This aggravates the fact Blocks don't guarantee Transaction parents are mentioned in the same Block or a previous Block as their children. If a parent is mentioned in the Block after the child, the parent will be finalized when the child is.

While that is closer to the end goal, it's also more erratic considering Transactions can now be finalized before they're out of Epochs, causing a double finalize.