Only verify Transactions which have a chance at becoming verified.
This adds a beaten bool to TransactionStatus. If a competitor is verified or a parent Transaction is beaten, this flag is set. Meros will not verify TXs with this flag set.

This is a bit naive. Meros should only verify along a tree. If Meros verifies X, which is a competitor to Y, Meros will still verify Z (which uses inputs from both), as long X and Y are both still in Epochs. This is because the X/Y loser isn't marked as beaten until finalization, unless one is already out of Epochs when the other is noticed. Therefore, Z is seen as first and has no parents beaten.

This quirky behavior is fine as it doesn't count as a MeritRemoval and defaulting solves any potential problems with... wasted Verifications.

So. This is perfectly performant, protocol compliant, and secure. There can be more done to improve it.

This also fixes a memory leak where Transactions which weren't verified because their parents weren't didn't have their Statuses cleared from the cache.