Delete the Transactions Branch TODO item.
This does leave a lot of tests undone.

I really want to revamp the Python suite before I do them, and don't want to do that on this branch. They have a lot of repeated code that makes it hard to understand what the tests actually do from a quick read and alter its behavior. The plan is to make a Live-version of Syncer which works of callbacks to trigger events at certain block heights.

Because there's a firm plan on what to do about the Python suite, because of how much that'll help with writing these new tests, because all Python tests pass and new Nim tests have been written... I'm confident with putting the tests off.