Change Client sync handling.
This PR removals one of the two status enum's we have in Meros: ClientState. It had the values of Ready and Syncing. It was used with each Client to define ourState and theirState.

ourState has been renamed syncLevels, which is an int.  Every call to startSyncing adds 1, and every call to stopSyncing removes 1. The actual SyncingOver message is only sent when syncLevels turns into 0.

This means syncBlock can call syncElements/syncTransaction, before closing it's own sync, in order to contain the entire sync process in a single Syncing/SyncingOver block.

That said, the current code doesn't provide an oppurtunity for sync levels to take effect. This is a technical edit for cleaner code and possibilities, not one which takes an effect.