Move the Merit tests over to Liver.
ChainAdvancementTest now executes a Syncer and a Liver, instead of using the RPC. SyncTest, which was the CALive test that also checked the Difficulty, has been renamed DifficultyTest and moved to Liver.

This move required storing all difficulties, not just the latest, in the Blockchain class. This change was needed at some point, and that point happened to be now.