Major Protocol Refactor.
I'm not happy with Meros. It pains me to admit it. I'm not happy with some aspects of the protocol, the lack of mainnet, the lack of other developers... most of which can be put on me.

I could have designed a better protocol in the first place. I could build a smaller protocol and expand with hard forks for a faster mainnet. I could have better marketing. I could take more freelance jobs and hire developers (although, that would slow down the rest of development).

This branch is named "No Consensus DAG", yet aims to solve a lot more. That's just the biggest change. It removes the Block Lattice we have for the Consensus DAG, adding a second layer on top of the Transactions DAG, shoe horning the rest in the Blockchain.

This uses less bandwidth, space, and CPU resources overall, yet does grow the Block size (so instead of X over time, it's Y at once, even though X < Y).  That said, based on how it can be improved via hard forks and the low usage planned at launch, I'm fine with that.

This also removes the nonce from the header, replacing it with a version field, and disables pool mining. Blocks are now mined to a single miner, who must do PoW, then sign the hash, then do more PoW. This makes it infeasible to work with a pool (who would have the Private Key).

Those are the note worthy changes. This definitely counts as a major refactor so it's subject to a lot of consideration and scrutiny.

If we make these changes, I'd like to do them before mainnet. Moving to a TX hash based indexing for Verifications, instead of Verifier/Nonce, will speed things up a lot. That said, if we put it off until after mainnet, we're stuck with maintaining the existing DAG forever, even if it's no longer updated. That's why this branch is here.