Distinguish between Element and BlockElement.
This change was needed to enable VerificationPackets as MeritRemoval causes. Elements no longer have holders. BlockElements, AKA Elements archived in Blocks, do, along with verifications (whereas VerificationPackets have a list).

Because Elements are no longer guaranteed to have a holder, MeritRemoval constructors require one be passed in. This also works to specify the problematic Merit Holder when VerificationPackets are the cause (although it's easy enough to check which Merit Holder is in both packets).

This also updates the Element equality operator to work with VerificationPackets.