Get the Consensus to compile.
This effectively gets the Consensus DAG to meet the new protocol, except for the lack of MeritRemovals. The one type we supported isn't part of the new protocol. While the new protocol adds a new type which is very similar, it requires code outside the scope of this branch.

Also comments Epochs out of CompareMerit as that broke the WorkingTests and corrects some of the GlobalFunctionBox.