Remove redundancies.
This will be a long commit.

The first redundancy was Blocks having transactions and packets. As every Verification Packet has their Transaction hash, and the packets were in the same order as the transactions, this was pointless.

The second was using two separate merkles for contents (contents and verifiers). Since the verifiers merkle has the fully serialized VerificationPacket, the verifiers merkle includes every Transaction. That said, to verify a Transaction inclusion, you now need the hash and the verifiers, instead of just the hash.

Since the verifiers is likely also necessary to anyone writing an SPV client, they're easy to get, and can't be forged without a hash collision, this is fine.

The final thing left was using two separate sketches. The first sketch was hashes, the second sketch was hashes + verifiers.

The first sketch is much more likely to resolve (hash or not when the second is hash + every verifier). That said, by removing the first sketch, we can double the capacity of the second sketch.

It's also difficult to handle the two separate sketches when the actual data in the BlockBody is unified. This provides a cleaner, safer, and more scalable solution.

As we now only have one sketch, it's no longer a generic.

Also updates the Eventual RPC docs with changes already made and defined in Current.

Also adds a function to check if a sketch salt causes a collision. This is utilized by the SerializeBlockBody and SerializeBlock tests (which naturally found a collision).

Also prepares Epochs to use the packets in the BlockBody objects, instead of separately passed in packets.