Update Python to use a single sketch/single contents merkle.
This also updates the Block classes to use sketches in the first place.

Also updates PyRight to be strict on every directory (as it wasn't on the Pylint directory) and ignore the Generate file (which imports Generators, yet doesn't use any properties from them, so PyRight has a false positive).

Also updates the BlankBlock vectors.