Clean ups.
The primary edit is renaming property names back from .txHash/.blockHash to .hash. I edited them to not conflict with hash, as it's a predefined builtin, yet it only conflicts with variables, not properties.

The hash variables are still txHash/blockHash.

Also removes DIFFICULTY_LEN, corrects the singular, improper, use of it, and removes an unused import.