Add a dedicated merkle to check the sketch against.
This reverts the contents merkle to be of the packets and elements.

Creates and utilizes BLOCK_HEADER_DATA_LEN, a length for the non-miner data in the BlockHeader. While the entire BlockHeader is data, the data in this section doesn't represent Block data.

Updates VCompeting's vector file to handle sorting rules.