Have publishBlock pass in a Sketcher.
If the Sketcher changes at all, so there's any differences, it will attempt to sync the missing data over the network. An RPC client is not a network node who can answer these queries.

Locking in the sketcher enables Verifications to be received while mining without having to download a new template (although these Verifications won't be present in the Block, of course).