Remove an invalid protocol rule we never implemented.
The rule was to use Data's Argon hash, not their hash, to index them. This was because we need the hash to not start with 16 0s (a near-impossible chance).The issue is anyone can set any PoW. It's not part of the signature to enable remote work.

Now that it's of the hash alone, there is a near-impossible risk some Data Transactions will not be possible without a buffer or padding.  This is acceptable due to the odds and ease of padding/buffering.