Port Meros to Milagro.
This swaps the groups used for Signatures/Public Keys, making Signatures 
48 bytes and Public Keys 96 bytes. As we've nicknamed Public Keys, this 
is optimal.

This branch also makes two additional edits, that are out of scope 
according to the branch-name. This moves Meros back to Nim's C backend, 
as:
1) We only needed the C++ backend for Chia.
2) Milagro didn't work with the C++ backend.
3) The only other factor was C++'s zero-cost Exceptions. Due to the 
heavily added cost when Exceptions are triggered, and the fact we use 
Exceptions frequently, this isn't a detriment.

This also turns BlockHeaders into refs. Public Keys/Signatures were 
randomly zero-ing and this change fixed those bugs.