Update documentation.
Removes the documentation on serializing BLS objects for the simple 
comment we follow ZCash's serialization standard.

Clarifies various aspects of the Consensus protocol.

Fixes a typo in the PythonTests' README.

Bridges the Debian and macOS compilation guide to an unified format.

Removes the out-of-date Windows compilation guide.

Moves the TODO to the Development docs folder.