Partially fix https://github.com/MerosCrypto/Meros/issues/109.
Instead of using either of the two described options, a time check is 
implemented. I don't know why it wasn't done this way in the first 
place.

Also tightens up timings and allows the client keep-alive timer to 
terminate.

The reason this is only a partial fix is because we still have one 
AsyncFD for all RPC clients (fine) and one AsyncFD for each Network 
client (not fine).