Reorg Block addition.
NetworkSync now verifies that if a MR is archived, it's the only 
archived Element for that holder when it checks the nonces. This means 
after NetworkSync, Block addition is confirmed.

Adding the VPs was moved into MainMerit and out of NetworkSync.

Adding the Elements was also moved into MainMerit out of NetworkSync. 
They were moved after archive as that is required for 
SendDifficulty/DataDifficulty/GasPrices. I believe MeritRemovals have to 
be handled much earlier.

Beyond questioning when to handle MeritRemovals, I also struggle to 
remember the purpose of process/postProcess as I know it didn't fix 
every issue we had. As of right now, there's no meaningful code in 
between the steps which validate it.