Fix https://github.com/MerosCrypto/Meros/issues/114.
As the issue stated, this was about the test reasoning for deleting 
elements not being equivalent to the filter's.

The test now deletes any difficulties without votes. The filter now 
deletes any difficulties without votes.

I tried preserving the list algo in SpamFilter. I truly did. I finally 
gave up on it for a seq when I had a list with votes:

1
1
2
1
1

and the code created two elements for the same difficulty, one not on 
the list. When deleting the element not on the list, it deleted the 
element on the list, yet used the junk from the element not on the list 
as `next`, forming:

1
1
2
0
0

This algorithm should be just as fast. Only difference (besides removal 
of 0-elements) is the use of a seq over a list. VotedDifficulty is still 
a ref object, it just no longer has prev/next.