Clean up the Database code.
The major change is the use of templates to generate keys. This puts 
every formula in one place, easily allowing checks for conflicting 
formulas, as well as edits, since one edit propogates to every usage in 
the codebase.

This also fixes existing conflicting formulas and updates how holders 
are handled in MeritDB.

Also fixes a bug in the Spendable test.