Bug fixes.
The primary bug fixes in this commit have to do with the Consensus 
Test's methodology. Because of the way it incremented/decremented Merit 
in increments of 25, the SpamFilter never realized when someone lost a 
vote (not to mention we always added 25 Merit, instead of subtracting 25 
Merit when Merit was lost).

We also failed to assign a holder to the send/data difficulties, so they 
were always of holder 0.

We also checked the close table when comparing Consensuses. This isn't 
possible unless we write extra checks to handle edge cases.

There was one actual bug in the SpamFilter, where if the same holder 
gains and loses Merit in the same Block, it could trigger a vote loss 
improperly.

I also simplified some of the if checks in the SpamFilter.handleBlock 
functions.