Remove the 25 Merit scaling from the ConsensusTest.
It was too problematic to work properly. Removing it required scaling 
everything by 25, which makes this test run much slower, but the 
pointless mining was removed (we never check against a difficulty) which 
helps.

This pointed out a couple of bugs in unverification which have been 
fixed (deleted unverified statuses from the cache/trued to unverify 
finalized transactions).

Also took the oppurtunity to apply a few optimizations.