Update the protocol to prevent unsyncable chains.
In order to verify a MeritRemoval caused by competing Verifications, the 
Transactions need to be checked for a common input. Meros does this by 
the Transactions DAG, which requires the Transactions to be unique or 
competing with something in Epochs. If a MeritReoval doesn't have its 
Verifications in a Block, Meros will refuse to add the Transaction, 
meaning it will refuse to verify the MeritRemoval.

This protocol change is to satisfy a software issue. That said, it's not 
a bad fix and should be beneficial for all implementations.