Update the protocol to prevent unsyncable chains.
As the previous commit titled this stated:

In order to verify a MeritRemoval caused by competing Verifications, the 
Transactions need to be checked for a common input. Meros does this by 
the Transactions DAG, which requires the Transactions to be unique or 
competing with something in Epochs. If a MeritReoval doesn't have its 
Verifications in a Block, Meros will refuse to add the Transaction, 
meaning it will refuse to verify the MeritRemoval.

It incorrectly declared this a software issue with a protocol fix. The 
protocol has the same requirements which caused the software issue.

This commit reverts the changes made by the original and allows 
Verifications to be of ANY parsable Transaction when present in a 
MeritRemoval.