Add support for peer finding and seed nodes.
Doesn't allow the same connection multiple times. Does allow if -rc 
(--allow-repeating-connections) is passed, an undocumented flag (as it's 
dangerous) meant solely for testing.

Also handles messages leading up to SyncingAcknowledged, instead of 
discarding them. This functionality, which helpful, was added because 
there's now a timed event which causes sync-mode to be activated (and 
therefore message discarding). If it was done improperly, there could be 
larger ramifications.