DT2 Fixes (#129)
* Add thread/async safety to addBlock*.

Does the same for addSignedMeritRemoval.

* Correct when the lock is released in addBlock.

* Add a missing check to addBlock.

Also replaces a panic with a return now that it's no longer a panic 
circumstance.

* Continue if we failed to acquire the lock.

* Fix https://github.com/MerosCrypto/Meros/issues/126.

* Fix lock acquisition for addBlock.

* getPeers RPC method.

* Fix publishBlock.

publishBlock directly calls addBlock yet addBlock requires the 
lockedBlock to be set. This modifies the external addBlock to acquire 
the lock and set it.

* Fix https://github.com/MerosCrypto/Meros/issues/126 (part 2).

* Bug fixes.

- Cleared lockedBlock when addBlockByHeader fails.

- Don't increment syncLevels when we failed to start syncing.

* Correct the starting point used to parse peers.

* Fix https://github.com/MerosCrypto/Meros/issues/128.

Also expands the test for #126.

* Fix syncing Blocks.

A previous fix broke syncing thanks to clearing the lockedBlock every 
time, not only on error.