Update the Python Tests.
Deletes HundredTwentySixTest as it relates to Sync Mode specifically.

Deletes Multiple Connections test as it's no longer testable. 127.0.0.1 
is allowed to connect multiple times and any equivalent IP isn't allowed 
to connect at all. It should be added back with a VLAN setup.

Adds a test for HundredTwentyFive, verifying it's fixed on any platform 
the tests are run on. Thanks to it, I'm willing to close #125 now.

Moves the protocol/network ID inside of the new MerosSocket class, 
removing the need to pass it on every instantiation.

Fixes multiple bugs in Meros, such as:

- Meros improperly responded to Syncing messages sent after the initial 
handshake.

- Meros only removed inactive peers at boot; it didn't set a timer to do 
it regularly.

- The function which reported if a Peer was closed actually reported if 
it was open.

- When Handshaking with a node about to timeout, Meros always used the 
live socket, even if it didn't exist.

- Meros returned instead of raising DataExists in one place, which could 
cause a rebroadcast of existing data. 

- Meros idented message creation, stopping certain message types from 
work.

- Meros didn't pad its serialization of Peers, breaking it.