Don't set closed sockets to null.
null is reserved for sockets which never existed. This prevents a lot of 
potential SIGSEGVs since sockets where used in multiple places yet a 
SocketError only notified one of the socket's invalidity.

Also provides `safeClose`, a safe way to close a socket. Just makes sure 
the socket isn't null and wraps the close call in a try/except block.