Fix Same Element Merit Removals.
Meros allowed these. They are completely invalid. They allow for any 
Merit Holder who created any Element or Verification to lose all their 
Merit.

Also includes a patch to #133 test. It's a failure test but was failing 
because it couldn't sync a referred to Transaction, not because the 
packet didn't include the MR's holder.

Also corrects a typo in MeritRemoval's class definition.