Don't route Consensus Transactions through the GlobalFunctionBox's getTransaction.
Doing so allowed Transactions to partially bypass the Transactions DAG, 
when they're only supposed to be able to for Merit Removals.

Also removes bloat from the Invalid Competing test, and unfortunately 
removes its Syncer. The Python code doesn't know not to verify 
Transactions which are solely part of MeritRemovals. We can overcome 
this with the Liver being a SuccessError but the Syncer doesn't have the 
neccessary callbacks.