Update the Send hash definition to remove any chance of collisions.
A Transaction with malformed inputs/outputs could have the same hash as 
a valid Send. It solely depends on how it was serialized before 
transmission. That said, if Meros ever wants to be able to identify 
invalid TXs it's already processed by hash alone, this is a key 
difference we need to be able to make.