Fix https://github.com/MerosCrypto/Meros/issues/119.
This also introduces a 'pending' hash set in preparation of 
https://github.com/MerosCrypto/Meros/issues/144. As chain 
reorganizations remove the validity of signature aggregates, we will 
need to preserve as many signatures as possible from unfinalized 
transactions (which when combined with checkpoints, should handle 
everything). That said, we currently use if the signature is present in 
the table to determine is a holder is pending.

We could just check the packet's holder seq, which is a O(n) operation, 
or for a small memory trade off, acquire a O(1) operation.