Initial Consensus Revert test.
It's based off the Transaction test, not the Consensus test. The 
Consensus test didn't generate a proper Transactions DAG which is a 
requirement for reversions.

This test doesn't generate a proper Consensus DAG, and there's still a 
lot of other checks required to implement and clean up to do.

Beyond that, it should be merged with the Transactions DB test. They 
share bones/the TX DAG generation + replay algo. It's really just 
porting the tracking vars and checks.

We could merge with the Consensus DB test as well except for two things:

- The Consensus DB has 1250 rounds which is unsuitable when we do a 
block by block revert. If we only revert the first 50 blocks one by one, 
then the rest in groups of tens... this could work.

- Beyond that, the generation requirements for the Consensus DAG are 
slightly different. This test requires some holders never have their 
elements archived, after a certain point. This isn't a deal breaker, 
just something to watch.

Also includes a minor bug fix to the revert algorithm.