Fix a false-positive for creating a MeritRemoval.
MeritRemovals can be created when one verifier verifies two Transactions 
which spend the same input. The Consensus DAG assumed 32 * 0 was a valid 
input, and since every initial Data spends it, anyone who verified 
multiple initial Datas got a MeritRemoval according to Meros.

This stops tracking any spenders for 32 * 0 when it's used for initial 
Datas. The (tx of Data) check is generally faster than the hash check, 
and then it wont check for 32 * 0 when the Transaction doesn't have 
inputs.

On this branch because it was holding the ConsensusRevert test.