Add a working replay algorithm.
Also uncomments the replay checks, along with the all-at-once revert 
check.

This test now solely needs to add Verifications of old Transactions, 
pending Verifications to the latest Transactions (to test the pending 
holders, pending packet, and signatures cache), Send and Data 
difficulties (to properly test SpamFilter reversions), and Merit 
Removals.

The first, third, and fourth's code can be ripped from the existing 
Consensus test. The second needs to be written from scratch. We do need 
to modify the Send/Data difficulty generation code to not archive 
certain Elements in order to test we purge pending Elements when a gap 
happens.

That doesn't test the competing/beaten fields. I'd like to write a 
separate unit test, or Python test, to handle those specific cases.