Several optimizations.
- Removes a holder. 0, 1, 2 are valid holders; 3 is reserved for testing 
pending Elements. This is to lower the amount of Blocks needed to ensure 
holders get enough Merit to vote.
- Makes the -time, +memory trade off with calculating what Verifications 
a Transaction should have.
- Changes the amount of Blocks to 155. Since holder #3 only ever 
receives 1 Merit, this should give at least 2 of the remaining holders 
enough Merit to vote on the difficulties.
- Changes the Block reverted to from 10 to 50. This means 105 Blocks are 
reverted, not 240.
- Reduces Transaction frequency to save on memory. The test was using 14 
GB of RAM previously. It now uses roughly 2 GB.