Correct the verified check which didn't handle a edge cases.
The specific edge case that was fixed in this is when a parent had 
pending verifications, and that caused the parent to become unexpectedly 
verified, and that caused all children above the threshold to become 
verified (unexpectedly).

Also corrects a bug in the discoverUnorderedTree algorithm where the 
root Transaction wasn't included.