Correct the assigned Epoch.
The Epoch was assigned to the height + 7, as that's what the Consensus 
test said it needed to be. Unfortunately, the Consensus test used an 
invalid height when registering the Transactions. Now, registering the 
Transaction is done with the proper height and the proper offset is 
applied.

When finalizing a Transaction, epoch == state.processedBlocks now holds 
true.