Fix https://github.com/MerosCrypto/Meros/issues/102.
The ConsensusRevert's verified check was failing, again, in relation to 
dead Merit this time. This fixes issue 102, a failure with how we handle 
dead Merit, in an attempt to fix tne check.

The new State algorithms require deadBlocks be greater than or equal to 
7 in order to not cause an IndexError. This is fine for our needs, yet 
if it wasn't, we could add an additional if check.

This doesn't fix the verified check failing, yet it does change which 
verified check is failing. Due to all the work put in, I want to 
preserve this work and it's required to be included in this branch due 
to how it affects the verified checks.