Redefine the Send/Data difficulty, remove GasPrice for GasDifficulty.
The Send/Data difficulties now work as the Blockchain difficulty is, 
except that it's decided by vote instead of a calculation.

GasPrice was removed because it goes against the ideals of Meros. A 
competent spam function shoud remove the needed for charging MR as a 
fee. Now that we have a competent spam function, GasDifficulty has been 
added.