Update Python's definition of Send and Data difficulties.
This slipped by me as all type checks passed with the SpamFilter. Unlike 
Nim, Python never actually implemented a SpamFilter (in a meaningful 
manner), so the checks passing were not certification the data types 
were updated.

This updates the Send and Data difficulties, as well as all relevant 
tests and the serialization/receiving code.