Bug fixes to the existing reorganization algorithm.
This doesn't extend it to work with a variable Merit Holders set, or 
work with a depth > 1. This solely fixes the algorithm so that when the 
Merit Holder set is constant, and the depth is 1, re-organizations are 
possible.

Also includes an optimization in MainMerit and a test to verify this 
functionality.