Remove toAddress.
Ever since utilizing Bech32 for the addresses, instead of our custom 
Base32 bit shift code, this has severely diminished functionality. Now 
that address types exist, this is ever further redundant.

We could extend it with the type of address to create, so we can support 
P2PK and P2PKH (when we eventually add P2PKH, if we add it), yet there's 
enough Bech32 libraries out there this isn't really needed.