Correct the heights at which the new RandomX key is set/updated.
RandomX recommends setting the upcoming key every 2048 blocks and 
switching over when it has a depth of 64. That said, this is based off 
the Monero Block time which is 1/5th of Meros's.

Updated 2048 to 384 (410 would've been more accurate yet I prefer 
rounder numbers, not to mention 2048 is equivalent to ~2.8 days which 
itself isn't round, showing that 2048 over 2160 was for the 
power-of-two) and 64 to 12 (13 would've been a bit more accurate, same 
reason as 384 > 410).